Fetch latest transactions from database
